requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn[standard]
joblib
python-multipart
Pillow
pytest
mlflow
streamlit
evidently
sniffio